[Fix] Silencing samples was slightly broken (http://bugs.openmpt.org/view.php?id=555). [Fix] XM Loader: Clear MIDI macros for XM files made with FT2 to prevent Zxx effects from doing anything. [Mod] OpenMPT: Version is now 1.23.04.03
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4178 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Ctrl_smp.cpp
@@ -2329,7 +2329,7 @@ void CCtrlSamples::OnSilence()
 				int delta = dest - base;
 				for (SmpLength i=0; i<len; i++)
 				{
-					int n = base + (int)(((LONGLONG)delta * (LONGLONG)i) / (len-1));
+					int n = base + (int)(((int64)delta * (int64)i) / int64(len - 1));
 					p[i] = (signed short)n;
 				}
 			} else
@@ -2340,7 +2340,7 @@ void CCtrlSamples::OnSilence()
 				int delta = dest - base;
 				for (SmpLength i=0; i<len; i++)
 				{
-					int n = base + (delta * i) / (len-1);
+					int n = base + (int)(((int64)delta * (int64)i) / int64(len - 1));
 					p[i] = (signed char)n;
 				}
 			}
